+++ MyMusic/MyMusic/FFPlayer/FFPlayer.h
@@ -28,7 +28,7 @@ signals:
     void stateChanged(PlayerState state); // 状态变化信号，用于控制播放器状态
     void initFinished();                // 初始化成功信号
     void PlayFinished();                // 播放完毕信号
-    void clockChanged(int clock);    // 时钟同步信号
+    void clockChanged(double clock);    // 时钟同步信号
 
 private:
     void fateError();
